app/core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Основной сервис для работы"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.118.0",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
